Moved all StripeSpecificationWrappers into the same sub-package (Review Feedback)
