+++ compatibility/compatibility-vec11to12/src/main/java/com/foursoft/harness/compatibility/vec11to12/Vec11XTo12XCompatibilityWrapper.java
@@ -32,6 +32,7 @@ import com.foursoft.harness.compatibility.core.mapping.ClassMapper;
 import com.foursoft.harness.compatibility.core.wrapper.CompatibilityWrapper;
 import com.foursoft.harness.compatibility.vec11to12.wrapper.vec11to12.*;
 import com.foursoft.harness.compatibility.vec11to12.wrapper.vec11to12.specification.*;
+import com.foursoft.harness.compatibility.vec11to12.wrapper.vec11to12.specification.wireprotection.Vec11To12StripeSpecificationWrapper;
 import com.foursoft.harness.compatibility.vec11to12.wrapper.vec12to11.*;
 import com.foursoft.harness.compatibility.vec11to12.wrapper.vec12to11.specification.*;
 import com.foursoft.harness.compatibility.vec11to12.wrapper.vec12to11.specification.terminal.Vec12To11PluggableTerminalSpecificationWrapper;
