+++ compatibility/compatibility-vec11to12/src/test/java/com/foursoft/harness/compatibility/vec11to12/wrapper/vec11to12/specification/wireprotection/Vec11To12StripeSpecificationWrapperTest.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.harness.compatibility.vec11to12.wrapper.vec11to12.specification;
+package com.foursoft.harness.compatibility.vec11to12.wrapper.vec11to12.specification.wireprotection;
 
 import com.foursoft.harness.compatibility.vec11to12.TestFiles;
 import com.foursoft.harness.compatibility.vec11to12.util.WrapperUtils;
